Procedures, predefined templates
HabboCMS will now run trough the script, check for errors and if no
error is displayed it'll print the output. However if not it'll display
a error message with debug information IF the script is in debug mode or
LOGGEDIN user has developer status. Procedures are quick scripts that
will be put together in main file to accomplish the goal. Each procedure
has it own task, no procedures contain more than one task.